move pushToOrFilterOutCollections to only user
